### 2025-09-30 #9118 @editorialbot
[PRE REVIEW]: CFRL: A Python library for counterfactually fair offline reinforcement learning via sequential data preprocessing
**Submitting author:** @JianhanZhang (<a href="http://orcid.org/0009-0000-1434-227X">Jianhan Zhang</a>)
**Repository:** https://github.com/JianhanZhang/CFRL
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.1
