### 2025-10-07 #9118 @JianhanZhang
Dear Potential Editors and Reviewers,

We just decided to change the name of our package from `CFRL` to `PyCFRL`. In this case, we made 3 major changes:

1. The title of the paper is now "PyCFRL: A Python library for counterfactually fair offline reinforcement learning via sequential data preprocessing".
2. The name of our Github repository is now "PyCFRL" rather than "CFRL". It can be accessed at either <https://github.com/JianhanZhang/CFRL> or <https://github.com/JianhanZhang/PyCFRL>.
3. We released a new version of the library on PyPI to reflect this name change. The most updated version is 0.3.0 at this point.

I am writing this update in the hope of mitigating potential confusion in the upcoming review process. Please let me know if you have any questions or concerns. I am sorry for the inconvenience, and thank you for your time and help!

Jianhan
