### 2025-10-28 #9118 @crvernon
@JianhanZhang - thank you for your submission to JOSS.  We are managing a large backlog of submissions and will get you a topic editor assigned as soon as possible.

In the meantime, please reduce the word count of your paper to <=1000 words.
