### 2025-12-23 #9689 @editorialbot
[REVIEW]: CFRL: A Python library for counterfactually fair offline reinforcement learning via sequential data preprocessing
**Submitting author:** @JianhanZhang (<a href="http://orcid.org/0009-0000-1434-227X">Jianhan Zhang</a>)
**Repository:** https://github.com/JianhanZhang/CFRL
**Branch with paper.md** (empty if default branch):
**Version:** 0.3.2
