### 2025-12-23 #9689 @rich2355
Hi @JianhanZhang, @lwu9, @christinamaher, it is so nice that we started the review right before the Christmas. Before we talk about the actual work, I want to firstly wish everyone a great holiday!

```@editorialbot generate my checklist```

**The JOSS review is different from most other journals.** Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention `openjournals/joss-reviews#9689` so that a link is created to this thread (and I can keep an eye on what is happening). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.

We aim for reviews to be completed within about 2-4 weeks. I will also remind the reviewers if there is no reviewing activity for a while (say, one week). However, as we are so close to the holidays, I may also need to take sometime off 😄. In addition, things are always evolving and please feel free to let me know if any of you requires some more time.

Now it is time when the reviewers provide comments for the authors to address. But please feel free to ping me (@rich2355) if you have any questions/concerns during the review process.

Best wishes,
RL
