### 2025-12-29 #9689 @JianhanZhang
> I created a virtual environment and followed the instructions in the authors' README by running the following in the command line:
>
> > $ pip install pycfrl
> > $ pip install pytest
> > $ python -m pytest
>
> It first gave me the errors as follows:
>
> > ModuleNotFoundError: No module named 'matplotlib'
> > ModuleNotFoundError: No module named 'seaborn'
>
> After I installed the above two libraries, it managed to run `pytest` and got the results below. Did the authors also get similar results, such as the number of skipped and warnings?
>
> > PyCFRL % python -m pytest
> > =========================================== test session starts ===========================================
> > platform darwin -- Python 3.11.14, pytest-9.0.2, pluggy-1.6.0
> > rootdir: /Users/lwu9/PycharmProjects/JOSS/PyCFRL
> > configfile: pyproject.toml
> > collected 74 items / 3 skipped
> > tests/test_agents.py ....ssss                                                                       [ 10%]
> > tests/test_evaluation.py ......ss..ss                                                               [ 27%]
> > tests/test_fqe.py ..ss                                                                              [ 32%]
> > tests/test_preprocessor_cf1_nn.py ............                                                      [ 48%]
> > tests/test_preprocessor_cf3_nn.py ............                                                      [ 64%]
> > tests/test_reader.py ..                                                                             [ 67%]
> > tests/test_real_data_analysis_integration.py ..ss                                                   [ 72%]
> > tests/test_simulated_env_nn.py ..........                                                           [ 86%]
> > tests/test_simulation_integration.py ss..                                                           [ 91%]
> > tests/test_synthetic_env.py ......                                                                  [100%]
> > ============================================ warnings summary =============================================
> > pycfrl/agents/agents.py:395: 5 warnings
> > pycfrl/fqe/fqe.py:439: 22 warnings
> > pycfrl/utils/base_models.py:373: 38 warnings
> > tests/test_agents.py: 1 warning
> > tests/test_evaluation.py: 4 warnings
> > tests/test_fqe.py: 2 warnings
> > tests/test_preprocessor_cf1_nn.py: 12 warnings
> > tests/test_preprocessor_cf3_nn.py: 20 warnings
> > tests/test_real_data_analysis_integration.py: 29 warnings
> > tests/test_simulated_env_nn.py: 10 warnings
> > tests/test_simulation_integration.py: 3 warnings
> > The fluctuation in the loss is not small enough in at least one of the final 10 epochs during neural network training
> > pycfrl/agents/agents.py:299: 5 warnings
> > tests/test_agents.py: 4 warnings
> > tests/test_real_data_analysis_integration.py: 2 warnings
> > tests/test_simulation_integration.py: 2 warnings
> > The fluctuation in the Q values is not small enough in at least one of the final 5 iterations during FQI training
> > pycfrl/fqe/fqe.py:460: 4 warnings
> > tests/test_evaluation.py: 2 warnings
> > tests/test_fqe.py: 2 warnings
> > tests/test_real_data_analysis_integration.py: 2 warnings
> > The fluctuation in the Q values is not small enough in at least one of the final 5 iterations during FQE training
> > -- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
> > ============================== 60 passed, 17 skipped, 169 warnings in 32.03s ==============================

Thank you so much for your comments! I just ran the tests locally and got similar results, with 60 passed, 17 skipped, and 167 warnings. I have pasted my results below:

```
================================================== test session starts ==================================================
platform win32 -- Python 3.10.8, pytest-8.4.0, pluggy-1.6.0
rootdir: E:\learning\university\MiSIL\CFRL Python Package\CFRL
configfile: pyproject.toml
plugins: Faker-37.1.0
collected 74 items / 3 skipped

tests\test_agents.py ....ssss                                                                                      [ 10%]
tests\test_evaluation.py ......ss..ss                                                                              [ 27%]
tests\test_fqe.py ..ss                                                                                             [ 32%]
tests\test_preprocessor_cf1_nn.py ............                                                                     [ 48%]
tests\test_preprocessor_cf3_nn.py ............                                                                     [ 64%]
tests\test_reader.py ..                                                                                            [ 67%]
tests\test_real_data_analysis_integration.py ..ss                                                                  [ 72%]
tests\test_simulated_env_nn.py ..........                                                                          [ 86%]
tests\test_simulation_integration.py ss..                                                                          [ 91%]
tests\test_synthetic_env.py ......                                                                                 [100%]

=================================================== warnings summary ====================================================
pycfrl\agents\agents.py:299: 5 warnings
tests/test_agents.py: 4 warnings
tests/test_real_data_analysis_integration.py: 2 warnings
tests/test_simulation_integration.py: 2 warnings

  The fluctuation in the Q values is not small enough in at least one of the final 5 iterations during FQI training

pycfrl\agents\agents.py:395: 5 warnings
pycfrl\fqe\fqe.py:439: 22 warnings
pycfrl\utils\base_models.py:373: 37 warnings
tests/test_agents.py: 1 warning
tests/test_evaluation.py: 4 warnings
tests/test_fqe.py: 2 warnings
tests/test_preprocessor_cf1_nn.py: 12 warnings
tests/test_preprocessor_cf3_nn.py: 20 warnings
tests/test_real_data_analysis_integration.py: 29 warnings
tests/test_simulated_env_nn.py: 10 warnings
tests/test_simulation_integration.py: 2 warnings

  The fluctuation in the loss is not small enough in at least one of the final 10 epochs during neural network training

pycfrl\fqe\fqe.py:460: 4 warnings
tests/test_evaluation.py: 2 warnings
tests/test_fqe.py: 2 warnings
pycfrl\fqe\fqe.py:460: 4 warnings
tests/test_evaluation.py: 2 warnings
tests/test_fqe.py: 2 warnings
tests/test_real_data_analysis_integration.py: 2 warnings

  The fluctuation in the Q values is not small enough in at least one of the final 5 iterations during FQE training

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
=============================== 60 passed, 17 skipped, 167 warnings in 146.16s (0:02:26) ================================
pycfrl\fqe\fqe.py:460: 4 warnings
tests/test_evaluation.py: 2 warnings
tests/test_fqe.py: 2 warnings
tests/test_real_data_analysis_integration.py: 2 warnings

pycfrl\fqe\fqe.py:460: 4 warnings
tests/test_evaluation.py: 2 warnings
tests/test_fqe.py: 2 warnings
tests/test_evaluation.py: 2 warnings
tests/test_fqe.py: 2 warnings
tests/test_fqe.py: 2 warnings
tests/test_real_data_analysis_integration.py: 2 warnings
  The fluctuation in the Q values is not small enough in at least one of the final 5 iterations during FQE training

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
=============================== 60 passed, 17 skipped, 167 warnings in 146.16s (0:02:26) ================================
```

We do expect that some tests will be skipped. During initial development, we provided options to use either neural networks or polynomial regression for the data preprocessor, the FQI agent, and the `SimulatedEnvironment`, and we wrote tests for both options. However, we later decided to only support neural networks for now. Consequently, all tests for the polynomial regression option are skipped.

The warnings are for potential nonconvergence in the training of neural networks, FQI, and FQE. In short, PyCFRL would monitor some metrics for convergence during training. If the changes in the metrics are not small enough during training, then a warning is raised. However, the warning is just a reminder of *potential* nonconvergence; it does not mean nonconvergence exists for sure. More details about the nonconvergence checks and warnings can be found [here](https://pycfrl-documentation.netlify.app/tutorials/common_issues).

Please let me know if you have any further questions or concerns. Thank you again for your time and comments!
