### 2025-12-31 #9689 @JianhanZhang
> Thanks to the author for the clarifications! I went ahead and continued to run the Jupyter notebook `examples/preprocessing_only_workflow.ipynb` and got 3 places different from the original notebook.
>
> 1. After I ran
>
> ```
> sp = SequentialPreprocessor(z_space=[[0], [1]],
>                             num_actions=2,
>                             cross_folds=1,
>                             mode='single',
>                             reg_model='nn')
> sp.train_preprocessor(zs=zs_train, xs=states_train, actions=actions_train, rewards=rewards_train)
> ```
>
> Some output is different, such as the last array in the output I got:
>
> ```
>  array([[ 0.01384445,  0.83824313,  0.14040247, ..., -1.66534019,
>          -0.44425263,  0.01813163],
>         [ 0.08313675,  2.03498991,  0.73120671, ...,  0.25546542,
>          -0.20163273, -0.17724276],
>         [ 1.83245792,  0.57598855,  1.04736866, ...,  0.26678068,
>           0.61744912, -0.3834829 ],
>         ...,
>         [ 1.89810574, -0.99365358,  0.07055229, ...,  0.16912496,
>          -0.7125387 , -0.07446041],
>         [ 0.51596592,  0.05168257, -0.29366556, ...,  0.44670743,
>           0.38861687, -0.28313736],
>         [ 1.42962013,  0.99360887,  0.41791617, ...,  0.68280647,
>           2.09172381, -0.13617618]], shape=(250, 10)))
> ```
>
> 2. After I ran
>
> ```
> states_tilde_cf5, rewards_tilde_cf5 = sp_cf5.train_preprocessor(zs=zs,
>                                                                 xs=states,
>                                                                 actions=actions,
>                                                                 rewards=rewards)
> ```
>
> I got
>
> ```
> 100%|██████████| 1000/1000 [00:11<00:00, 85.66it/s]
> 100%|██████████| 1000/1000 [00:11<00:00, 85.89it/s]
>
> The fluctuation in the loss is not small enough in at least one of the final 10 epochs during neural network training
> 100%|██████████| 1000/1000 [00:11<00:00, 88.48it/s]
>
> The fluctuation in the loss is not small enough in at least one of the final 10 epochs during neural network training
> 100%|██████████| 1000/1000 [00:11<00:00, 86.80it/s]
>
> The fluctuation in the loss is not small enough in at least one of the final 10 epochs during neural network training
> 100%|██████████| 1000/1000 [00:11<00:00, 86.64it/s]
> ```
>
> What does this mean "The fluctuation in the loss is not small enough in at least one of the final 10 epochs during neural network training"? And the original notebook only got one of those.
>
> 3. And for the output of `preprocessed_trajectory_cf5 `in the last cell, I got
>
> ```
> ID | time_step | z1 | action | reward | state1 | state2
> -- | -- | -- | -- | -- | -- | --
> 1.0 | 1.0 | 0.0 | NaN | NaN | 1.324345 | 2.287555
> 1.0 | 2.0 | 0.0 | 1.0 | 1.987784 | -0.813722 | 1.719199
> 1.0 | 3.0 | 0.0 | 1.0 | 0.581407 | -0.526683 | 2.641088
> 1.0 | 4.0 | 0.0 | 1.0 | 1.276694 | -0.464447 | 2.898451
> 1.0 | 5.0 | 0.0 | 1.0 | 1.452953 | -2.075518 | 1.529634
> ... | ... | ... | ... | ... | ... | ...
> 500.0 | 7.0 | 1.0 | 1.0 | -2.723090 | -2.179637 | -0.941954
> 500.0 | 8.0 | 1.0 | 1.0 | -1.682228 | -4.070047 | -2.536595
> 500.0 | 9.0 | 1.0 | 0.0 | -1.292153 | -1.904209 | -0.946557
> 500.0 | 10.0 | 1.0 | 0.0 | -0.545501 | -1.577761 | -0.709017
> 500.0 | 11.0 | 1.0 | 0.0 | -0.452039 | -1.684842 | -0.761890
> ```
>
> I wonder where those differences come from since we already set the random seeds to the same?

Thank you so much for bringing this problem to our attention!

Regarding the output differences, may I confirm one thing before I get to my conclusions? When you run this notebook twice *on your own device*, do you get the same output for both runs?

Regarding the warning, the warning message "The fluctuation in the loss is not small enough in at least one of the final 10 epochs during neural network training" is part of the nonconvergence check that we discussed earlier. In particular, the `SequentialPreprocessor` uses a neural network to estimate the transition kernel of the underlying environment. In our case, the warning is raised because the fluctuation in the loss is not small enough in the final 10 epochs of neural network training, which might indicate that the neural network has not converged yet.

However, this message does not mean that nonconvergence definitely exists or that the model is necessarily bad. Also, users can indeed customize how much fluctuation is considered "small", as well as how many final epochs to monitor (it is currently 10 in the tutorial). Thus, whether the warning is raised also depends on these user-specified parameters.

Please let me know if you have any further questions. Thank you again for your time and careful review!
