### 2026-01-05 #9689 @lwu9
Thanks to the author for the clarification. I ran that notebook twice on my device and got the same output for both runs. Why does a different device produce different outputs? Would it lead to different policies after learning?
