### 2026-01-05 #9689 @rich2355
I may add my two cents here but feel free to take or leave: I agree with the author that same random seed does not guarantee same output results due to the difference in the computing systems, but the results should not differ "so much" that the final decision based on the analytical results becomes unstable. I will leave the reviewer to decide on the stability of the algorithm.
