### 2026-02-02 #9689 @lwu9
Thanks to the authors for the explanations and for their efforts to check reproducibility across different platforms.
I kept running the notebook examples/synthetic_data_workflow.ipynb, and got the outputs of ` value` is `np.float64(9.27624836470061)` and `cf_metric` is `np.float64(0.46799999999999997)`. I guess they are still similar to what the original output was in the notebook. I wonder whether these differences are expected?
Another small thing is when I ran `from examples.baseline_agents import RandomAgent`, I got `ModuleNotFoundError: No module named 'examples'`, then it works after I changed it to
```
import sys
import os
# Add the parent directory to the path to enable imports from examples
sys.path.insert(0, os.path.dirname(os.path.abspath('.')))
from baseline_agents import RandomAgent
```
