### 2026-02-02 #9689 @lwu9
I run the notebook examples/real_data_workflow.ipynb, and got different but similar to the original output
1. the `value`: `np.float32(7.5081964`), `cf_metric`: `np.float64(0.03363636363636364)`, which has 12.12% relative difference from the original output (0.0299999999),` value_random`: `np.float32(-1.1875486)`, etc. Are these differences expected?
2. In [39]
```
from cfrl.preprocessor import Preprocessor
from cfrl.agents import Agent
```

should be changed to:
```
from pycfrl.preprocessor import Preprocessor
from pycfrl.agents import Agent
```
