### 2026-02-05 #9689 @JianhanZhang
> Thanks to the authors for the explanations and for their efforts to check reproducibility across different platforms. I kept running the notebook examples/synthetic_data_workflow.ipynb, and got the outputs of ` value` is `np.float64(9.27624836470061)` and `cf_metric` is `np.float64(0.46799999999999997)`. I guess they are still similar to what the original output was in the notebook. I wonder whether these differences are expected? Another small thing is when I ran `from examples.baseline_agents import RandomAgent`, I got `ModuleNotFoundError: No module named 'examples'`, then it works after I changed it to
>
> ```
> import sys
> import os
> # Add the parent directory to the path to enable imports from examples
> sys.path.insert(0, os.path.dirname(os.path.abspath('.')))
> from baseline_agents import RandomAgent
> ```

Thank you very much for your comments!

Regarding the first question, I again ran the script on Google Colab and got slightly different results from the results I got from my local machine. Similar to the previous cases, I highly suspect these differences are due to differences in the computing environment/system, and I think such differences should be expected.

Regarding the second question, I think you are correct. Since ``examples`` is not a module of PyCFRL, we cannot import `RandomAgent` by running ``from examples.baseline_agents import RandomAgent``. To fix this, I have updated the code tutorial, and in the updated tutorial, the ``RandomAgent`` class is now defined locally rather than imported. I apologize for the confusion and appreciate the catch.

Please let me know if you have any further questions or comments. Thank you again for your comments!
