### 2026-02-05 #9689 @JianhanZhang
> I run the notebook examples/real_data_workflow.ipynb, and got different but similar to the original output
>
> 1. the `value`: `np.float32(7.5081964`), `cf_metric`: `np.float64(0.03363636363636364)`, which has 12.12% relative difference from the original output (0.0299999999),` value_random`: `np.float32(-1.1875486)`, etc. Are these differences expected?
> 2. In [39]
>
> ```
> from cfrl.preprocessor import Preprocessor
> from cfrl.agents import Agent
> ```
>
> should be changed to:
>
> ```
> from pycfrl.preprocessor import Preprocessor
> from pycfrl.agents import Agent
> ```

Thanks again for these insightful comments!

**For the first question**, I also got slightly different outputs when running on Google Colab vs. running on my local machine. Similar to the previous cases, I think these differences are likely due to differences in the computing environment, and they should be expected.

Specifically, I am hoping to discuss a little more about the CF metric that has a 12.12% relative difference. In the example given in `real_data_workflow.ipynb`, the CF metric is calculated as

$$\text{CF metric} = \frac{\text{Number of counterfactually unfair decisions across all time steps and all individuals}}{\text{Total number of decisions across all time steps and all individuals}}.$$

In the dataset used in the tutorial, there are a total of 5000 decisions across all time steps and all individuals (i.e. the denominator in the above formula is 5000). In this case, a CF metric of approximately 0.029999 means that there are around $5000 \times 0.029999 \approx 150$ counterfactually unfair decisions, while a CF metric of approximately 0.033636 means that there are around $5000 \times 0.033636 \approx 168$ counterfactually unfair decisions. I feel it could be argued that this difference is still not significant, considering that there are a total of 5000 decisions.

**For the second question**, I have fixed the typo and pushed the corrected tutorials to the project repo. Thanks for pointing this out.

Please let me know if any further clarifications are needed. Thanks again for your time and help!
