### 2026-03-03 #9689 @christinamaher
**Summary:** PyCFRL represents an important and timely contribution to the field of counterfactual fairness in reinforcement learning. The software addresses a meaningful methodological gap and is accompanied by thorough documentation within the repository. The integration of simulation, evaluation, and policy learning workflows is well organized and reflects careful engineering.

**Installation and Testing**: The package installed successfully. As mentioned by @lwu9 , during testing I encountered missing dependencies (`matplotlib` and `_seaborn_`) that were not automatically installed with the package. After manually installing these libraries, I was able to run the full test suite. All tests executed successfully (60 tests passed; 17 skipped). Also similar to @lwu9, warnings were raised during neural network and FQI/FQE convergence checks, but these did not affect overall test success.

**Documentation**: The repository includes thorough and helpful documentation, and the testing infrastructure is robust. That said, I was unable to locate community contribution or troubleshooting guidelines.

**Paper and Presentation**: The accompanying manuscript is clear and well written. However, I have a few suggestions that may strengthen the presentation:

_**1. Workflow Visualization:**_ The summary section could benefit from inclusion of the schematic of the full PyCFRL workflow which appears in the repository.
**_2. Empirical Illustration of Bias Reduction:_** The paper would be strengthened by including a visual demonstration of empirical results (e.g., quantified reduction in bias when implementing the pipeline, a visual comparison of counterfactual fairness metrics, trained vs. untrained policy fairness comparison). A concrete example showing fairness improvement would make the methodological contribution more immediately interpretable and impactful.

PyCFRL is a valuable and well-engineered contribution to counterfactual fairness research in RL. The software installs cleanly (with minor dependency adjustments), the test suite passes, and the documentation is robust. Addressing minor dependency issues, adding contribution guidelines, and strengthening the empirical illustration in the manuscript would further improve an already strong submission!
