### 2026-04-17 #9689 @JianhanZhang
Hello @rich2355 and @christinamaher,

I have been occupied recently, and I want to apologize for the delay. Following the comments, I just added the workflow chart to the paper.

We are still working on community contribution or troubleshooting guidelines and the empirical results. In particular, there remain some details about the guidelines that need to be figured out. We are aiming to finish these two points by the end of May.

Please let me know if you have any questions or concerns, or if you feel an alternative timeline would be better. This review process has significantly improved our work. Thank you again for your time, patience, and comments!
