### 2026-04-18 #9689 @rich2355
> Hello [@rich2355](https://github.com/rich2355) and [@christinamaher](https://github.com/christinamaher),
>
> I have been occupied recently, and I want to apologize for the delay. Following the comments, I just added the workflow chart to the paper.
>
> We are still working on community contribution or troubleshooting guidelines and the empirical results. In particular, there remain some details about the guidelines that need to be figured out. We are aiming to finish these two points by the end of May.
>
> Please let me know if you have any questions or concerns, or if you feel an alternative timeline would be better. This review process has significantly improved our work. Thank you again for your time, patience, and comments!

Yes, I completely understand that things could happen. Glad that you come back and say hi to us!

I am happy with your timeline, @christinamaher this submission will be "paused" until the author @JianhanZhang is back. Please feel free to let us know if you have questions or concerns about this. Thanks!
