### 2026-08-16 #9689 @JianhanZhang
> Hi [@JianhanZhang](https://github.com/JianhanZhang), please check the missing DOIs (which must be corrected) and SKIP DOIs (which can be kept unaddressed if there is no existing DOI for a given ref indeed). In addition, please give me the DOI for your repo (by archiving the release) and the version number. Make sure that everything in your archive is the same with what shown in the github repo and paper (e.g. author list). Thanks!

Thank you @rich2355 again for your coordination! We have checked the points you mentioned.

**Missing DOIs:** I added the DOI for the "Neural Fitted Q Iteration" paper. For the "Gymnasium" paper, the DOI found by editorialbot seems to be valid. However, this DOI seems to correspond to another version of the paper, while the Gymnasium package recommends citing the arXiv version of the paper (please see [here](https://github.com/farama-foundation/gymnasium)), which does not come with a DOI. Therefore, I wonder whether it might be OK to follow the package authors' suggestion and cite the arXiv version with no DOI.

**SKIP DOIs:** I checked the references that fall into the "SKIP DOIs" category. I agree that all of these references seem to have no DOIs.

**Archive of release:** We archived the release on Zenodo. The Zenodo DOI is [10.5281/zenodo.21914207](https://doi.org/10.5281/zenodo.21914207). The version number is 0.3.2.

Please let us know if there are any further problems or concerns, or if any additional actions are needed on our side. Thank you again for your time and help!
