### 2026-08-19 #9689 @rich2355
Hi @JianhanZhang, editorialbot implies an issue in your reference file. Your `bib` file wrongly put a DOI into the `url` parameter for the `gym` reference:
`@article{towers2024gymnasium,
  title={Gymnasium: A Standard Interface for Reinforcement Learning Environments},
  author={Towers, Mark and Kwiatkowski, Ariel and Terry, Jordan and Balis, John U and De Cola, Gianluca and Deleu, Tristan and Goul{\~a}o, Manuel and Kallinteris, Andreas and Krimmel, Markus and KG, Arjun and others},
  journal={arXiv preprint arXiv:2407.17032},
  year={2024},
  url = {https://doi.org/10.48550/arXiv.2407.17032},
}`

Please fix this by putting this one as a DOI and let me know whether this works. Thanks!
