### 2026-08-29 #9689 @JianhanZhang
> [@editorialbot](https://github.com/editorialbot) set https://doi.org/10.5281/zenodo.21914208 as archive
>
> [@JianhanZhang](https://github.com/JianhanZhang) - I just need you to address the following before we move to publish.
>
> * [ ]  Why was a GitHub release not made?  I see your 0.3.2 push to PyPI but no accompanying release on GitHub.[ ]  The archive DOI is mismatched with what is declared in this thread.  I'll reset that as a part of this comment above.[ ]  The title of your Zenodo archive must match the title of your paper exactly, it currently does not.  Please edit the metadata of your Zenodo archive to reflect the title of your paper.[ ]  The author list and order in your Zenodo archive must match the author list and order in your paper exactly.
>
> In your paper:
>
> * [ ]  LINE 116: Capitalize the “Q” in the reference title “Neural fitted q iteration” as “Neural fitted Q iteration” to preserve the mathematical capitalization.
>
> Let me know when these have been addressed. Thank you.

Thank you very much for bringing these problems to our attention.  I have fixed the problems you mentioned above.

Please let us know in case there are any further problems we need to address. Thanks again for your time and help!
