### 2026-08-31 #9689 @JianhanZhang
> 🥳 Congratulations on your new publication [@JianhanZhang](https://github.com/JianhanZhang) ! Many thanks to [@rich2355](https://github.com/rich2355) for editing and [@lwu9](https://github.com/lwu9) and [@christinamaher](https://github.com/christinamaher) for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
>
> Please consider becoming a reviewer for JOSS if you are not already: https://reviewers.joss.theoj.org/join

Thank you very much @crvernon for your editorial help! Also thank you very much @rich2355, @lwu9, and @christinamaher for your coordination and review! Your comments and efforts have significantly improved our work!
